app/requirements.txt
fastapi[all]
requests
python-dotenv
sqlalchemy
psycopg2-binary